deleting uneeded files and moving service files
