+++ my-app/src/components/admin/CompletedRoomDetails.jsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
-import { getRoom } from "../../../services/roomsService";
-import { getUser } from "../../../services/usersService";
-import { getUsersSurvey } from "../../../services/surveyService";
+import { getRoom } from "../../services/roomsService";
+import { getUser } from "../../services/usersService";
+import { getUsersSurvey } from "../../services/surveyService";
 import { buildConversation } from "../survey/surveyUtils";
 import games from "../../gameLoader";
 import './admin.css';
